feat: Code restructure prior to adding new "Calendar" command
