+++ bin/remind
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('../dist/index.js');
+require('../dist/main.remind.js');
